Move commands-specifics to a compiler pass in FWB
