+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -67,6 +67,7 @@
         "symfony/asset": "<3.4",
         "symfony/console": "<3.4",
         "symfony/form": "<4.1",
+        "symfony/messenger": "<4.1.5",
         "symfony/property-info": "<3.4",
         "symfony/serializer": "<4.1",
         "symfony/stopwatch": "<3.4",
